BACKPORT: compiler, clang: suppress warning for unused static inline functions
GCC explicitly does not warn for unused static inline functions for
-Wunused-function.  The manual states:

Warn whenever a static function is declared but not defined or a
non-inline static function is unused.

Clang does warn for static inline functions that are unused.

It turns out that suppressing the warnings avoids potentially
complex #ifdef directives, which also reduces LOC.

Suppress the warning for clang.

Signed-off-by: David Rientjes <rientjes@google.com>
Signed-off-by: Linus Torvalds <torvalds@linux-foundation.org>

Bug: 32455818
BUg: 33244601
Change-Id: I31a7ff4b828538a66d439203be8965c3d53e6763
Signed-off-by: Nick Desaulniers <ndesaulniers@google.com>
Signed-off-by: Joe Maples <joe@frap129.org>